backend/requirements.txt
flask==2.3.3
python-dotenv==1.0.0
flask-cors==4.0.0
cryptography==41.0.4
iqoption==6.8.9.1
